Switch to use heira and packaged puppet
So HMRC, IDA and ERTP have used this pattern successfully for a while.
We package up puppet into a tarfile, complete with all dependencies
(Except ruby of course), and at vagrant up time we simply extract and
run the zip.
Actually it's a bit more complex than this because we do the package
step at vagrant up time, but in the future we could do that on a CI box
if we want.

We're also now using Hiera to control puppet, which will start to make
life easier.  I've left in the requirement for hiera-gpg but we're not
using it yet because we won't need it until we start actually
provisioning machines.